hllm/scripts/requirements.txt
requests
beautifulsoup4
openai
pillow